y2038: convert posix clock syscalls
This patch does a lot of things at once, and needs to be split
up into several small steps:

- split out base function from non-compat syscalls
- convert all nanosleep syscall restart handling to use __kernel_timespec
- make compat syscalls use the common base functions
- change normal syscalls to use __kernel_timespec

Most syscalls here are done in a simple way, the main complications
is from cleaning up the compat handling to not take an extra
step of get_fs()/set_fs(), which we would not actually have to do.

The nanosleep handling is still very ugly, maybe we can come up with
a better implementation for that in the process as well.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>