[NOT YET] y2038: make CONFIG_COMPAT_TIME optional
This makes CONFIG_COMPAT_TIME user selectable, with the default remaining
enabled so we don't break existing user space. The hacks to define the
__kernel_timespec, __kernel_stat, __kernel_rusage and __kernel_timex
structures to their legacy counterparts are removed now, in order to
keep the new behavior of the system calls.

The idea behind turning off this option is that users that want to build
long-term supported kernels can disable all code that is known to stop
working in 2038, so they can debug the problems in user space today instead
of guessing what might happen.

WARNING: This patch must not be applied before all 32-bit architectures
are converted to select CONFIG_ARCH_HAS_COMPAT_TIME and provide the new
system call entry points.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>