y2038: remove unneeded ipc uapi header files
Eight architectures contain copies of the original i386
ipcbuf/msgbuf/sembuf/shmbuf header files, which are all identical
to the version in uapi/asm-generic.

This patch removes the files and replaces them with 'generic-y'
statements, to avoid having to modify each copy when we extend
sysvipc to deal with 64-bit time_t.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>