y2038: x86: extend sysvipc data structures
This extends the x86 copy of the sysvipc data structures
to deal with 32-bit user space that has 64-bit time_t
and wants to see timestamps beyond 2038.

Fortunately, x86 has padding for this purpose in all the
data structures, so we can just add extra fields. For consistency
with big-endian architectures that copied the definitions from
x86 without moving around that padding, we however don't just
us a 64-bit value here, but instead use two unsigned 32-bit
fields, and expect the libc to provide a correct definition
to user space in a way that works across all architectures.

Interestingly, x86-64 is the only 64-bit architecture that
uses padding for 128-bit time_t, which seems to be have been
a mistake in the original x86_64 port. All other 64-bit
architectures have dropped those padding words.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>