y2038: asm-generic: extend sysvipc data structures
Most architectures now use the asm-generic copy of the sysvipc
data structures (msqid64_ds, semid64_ds, shmid64_ds), which use
32-bit __kernel_time_t on 32-bit architectures but have padding
behind them to allow extending the type to 64-bit.

Unfortunately, that fails on all big-endian architectures, which
have the padding on the wrong side. As so many of them get it wrong,
we decided to not bother even trying to fix it up when we introduced
the asm-generic copy. Instead we always use the padding word now
to provide the upper 32 bits of the seconds value, regardless of the
endianess.

A libc implementation on a typical big-endian system can deal with
this by providing its own copy of the structure definition to user
space, and swapping the two 32-bit words before returning from
the semctl/shmctl/msgctl system calls.

ARM, s/390 and Tile are architectures that use these generic headers
and also provide support for compat mode on 64-bit kernels, so
we adapt their copies here as well.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>