y2038: extend remaining sysvipc data structures
parisc, powerpc, sparc and xtensa all use incompatible variations
of the generic sysvipc data structures, intended to have the
padding moved around so they can deal with big-endian 32-bit
user space that has 64-bit time_t.

parisc and sparc actually succeeded in defining this right, but
as everyone else got it wrong, we just use the same hack
everywhere.

powerpc has the same definition as parisc and sparc, but now also
supports little-endian mode, which gets it wrong. xtensa tried
to fix it up many years ago but only got it right for two of the
three structures, while their struct shmid64_ds ended up being
defined in two identical copies, and the big-endian one is wrong.

In the end, we just take the same approach here that we have for
the asm-generic header and add separate 32-bit fields for the
upper halves of the timestamps, and let libc deal with the mess
in user space.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>