y2038: use __kernel_stat for sys_newstat syscalls
This moves over the sys_new{stat,lstat,fstat,fstatat} system calls
to use the newly introduced __kernel_stat. On all 64-bit architectures,
the layout is the same as 'struct stat', so nothing changes.

On 32-bit architectures, we use a trick to '#define __kernel_stat stat'
as long as CONFIG_COMPAT_TIME is not set, and hide the new definitions
from the kernel as long as this is the case, so the behavior does not
change until we change over the architectures individually to use
compat_sys_stat() in place of sys_stat() so we can reuse the sys_stat()
implementation for the new structure.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>