[SUBMITTED 8378] ARM: remove __bad_xchg() definition
We only define xchg() for 1-byte and 4-byte types, and call
__bad_xchg() otherwise, which results in a runtime error.

For cmpxchg(), we use the same method, but we intentionally
do not have a __bad_cmpxchg() definition and instead rely on
getting a compile-time error, which catches the problem much
earlier.

This removes the definition of __bad_xchg() and leaves the
caller, to use the same trick as on cmpxchg().

Signed-off-by: Arnd Bergmann <arnd@arndb.de>