posix-timers: Implement y2038 safe timer_get64() callback
The timer_get() callback in struct k_clock is not year 2038 safe on
32bit systems.

To address this implement a new callback timer_get64() which uses
struct itimerspec64 along with a default implementation which is a
wrapper for the existing timer_get() callback. The default callback
is installed at registration time for all posix clocks which are not
yet converted to timer_get64() and will be removed once this is
completed.

Use the new callback in __timer_gettime().

Signed-off-by: Baolin Wang <baolin.wang@linaro.org>

Conflicts:
	kernel/time/posix-timers.c