posix-timers: Implement y2038 safe clock_get64() callback
The clock_get() callback in struct k_clock is not year 2038 safe on
32bit systems.

To address this implement a new callback clock_get64() which uses
struct timespec64 along with a default implementation which is a
wrapper for the existing clock_get() callback. The default callback
is installed at registration time for all posix clocks which are not
yet converted to clock_get64() and will be removed once this is
completed.

Use the new callback in __clock_gettime().

Signed-off-by: Baolin Wang <baolin.wang@linaro.org>

Conflicts:
	include/linux/posix-timers.h
	kernel/time/posix-timers.c