posix-timers: Implement y2038 safe clock_getres64() callback
The clock_getres() callback in struct k_clock is not year 2038 safe on
32bit systems.

To address this implement a new callback clock_getres64() which uses
struct timespec64 along with a default implementation which is a
wrapper for the existing clock_getres() callback. The default callback
is installed at registration time for all posix clocks which are not
yet converted to clock_getres64() and will be removed once this is
completed.

Use the new callback in __clock_getres().

Signed-off-by: Baolin Wang <baolin.wang@linaro.org>

Conflicts:
	kernel/time/posix-timers.c