tsacct: avoid stupid timeval conversion
The tsacct code converts a cputime value to jiffies, then to
timeval, and finally to microseconds. This is pointless because
we have a cputime_to_usecs function that is even used in the
same file.

Worse, it breaks when the total time is more than the maximum
for 'jiffies', which can be as low as 49 days.

This changes the code to use the helper directly.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>