kdb: use avoid open-coded time_to_tm
kdb uses an open-coded version of time_to_tm() for printing the current
time, and suffers from the time_t overflow in 2038 on 32-bit architectures.

Instead of fixing the private implementation, this changes the code to
use the generic time64_to_tm that gets it right.

In order to read the current time without blocking on the xtime lock,
the code currently calls __current_kernel_time, but that returns a
32-bit seconds value, and we have to change that as well.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>