[experimental] ALSA: use private type for timestamp UAPI
This converts the use of 'struct timespec' in ALSA to a newly
introduced 'struct snd_timespec' type with the same definition,
in order to make the API independent of the user space definition
of 'struct timespec' that will change to use 64-bit members
for 32-bit systems in the future.

This is safe for monotonic timestamps, but not necessarily for
real time. If we can ensure that all user space uses monotonic
times by 2038, this approach will be the simplest solution,
otherwise we need to do something more sophisticated and
change the subsystem to handle both APIs (32-bit and 64-bit
time_t) transparently, based on the ioctl command codes that
are passed to the kernels.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>