[SUBMITTED] [media] v4l2: introduce v4l2_timeval
The v4l2 API uses a 'struct timeval' to communicate time stamps to user
space. This is broken on 32-bit architectures as soon as we have a C library
that defines time_t as 64 bit, which then changes the structure layout of
struct v4l2_buffer.

Since existing user space source code relies on the type to be 'struct
timeval' and we want to preserve compile-time compatibility when moving
to a new libc, we cannot make user-visible changes to the header file.

In this patch, we change the type of the timestamp to 'struct v4l2_timeval'
as a preparation for a follow-up that adds API compatiblity for both
32-bit and 64-bit time_t. This patch should have no impact on generated
code in either user space or kernel.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>