ceph: fix message length computation
create_request_message() computes the maximum length of a message,
but uses the wrong type for the time stamp: sizeof(struct timespec)
may be 8 or 16 depending on the architecture, while sizeof(struct
ceph_timespec) is always 8, and that is what gets put into the
message.

Found while auditing the uses of timespec for y2038 problems.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Fixes: b8e69066d8af ("ceph: include time stamp in every MDS request")