staging/lustre: remove CFS_TIME_T definition
The CFS_TIME_T macro serves no real purpose as we stopped using time_t
and changed over to time64_t, so we can remove the last remaining uses
of this.

Two uses of this macro are incorrect and refer to jiffies values
rather than time_t, and one refers to an inode timespec that gets
changed separately.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>