staging/lustre: use jiffies for lp_last_query times
The recently introduced lnet_peer_set_alive() function uses
get_seconds() to read the current time into a shared variable,
but all other uses of that variable compare it to jiffies values.

This changes the current use to jiffies as well for consistency.

This likely changes behavior, and should be reviewed by the folks
that reviewed the original patch that introduced the unusual behavior
and was itself a bug fix.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Fixes: af3fa7c71bf ("staging/lustre/lnet: peer aliveness status and NI status")
Cc: Liang Zhen <liang.zhen@intel.com>
Cc: James Simmons <uja.ornl@gmail.com>
Cc: Isaac Huang <he.huang@intel.com>
Cc: Oleg Drokin <oleg.drokin@intel.com>