y2038: introduce time64_to_tm()
The time_to_tm() function converts a time_t into a struct tm, and suffers
from the time_t overflow on 32-bit architectures in 2038.

This changes the implementation to use a time64_t instead, and rename
the function name to time64_to_tm(), turning time_to_tm() into a static
inline helper. The main purpose of the renaming rather than changing the
prototype is to catch all users of the the function that pass a 32-bit
number into it, so we can convert them as well.

To make the code more efficient on 32-bit systems, this implementation
avoids the expensive 'rem = do_div(totalsecs, SECS_PER_DAY);' calculation
and instead relies on the fact that the low seven bits of SECS_PER_DAY
are zero, and that we can calculate times until year 19378 with just
39 bits worth of seconds.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>