fat: use time64_to_tm
The time_to_tm function is not y2038-safe on 32-bit architectures, so
this changes the FAT file system over to use the new time64_to_tm
helper.

It's probably best to wait with this patch until we can pass 64-bit
timestamps in the inode as well, I'm just documenting it here for
the moment, while we wait for the file system patches to get rebased.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>