[WRONG] input: ams-delta: try to fix section mismatch (probably wrong)
WARNING: drivers/built-in.o(.exit.text+0x21cc): Section mismatch in reference from the function ams_delta
_serio_exit() to the (unknown reference) .init.rodata:(unknown)
The function __exit ams_delta_serio_exit() references
a (unknown reference) __initconst (unknown).
This is often seen when error handling in the exit function
uses functionality in the init path.
The fix is often to remove the __initconst annotation of
(unknown) so it may be used outside an init section.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>