[RESEND] ncpfs: avoid unused variable warning
When CONFIG_NCPFS_NLS is not set, we get a harmless warning
about an unused local variable, because that is only passed
into a macro that does not evaluate its argument:

fs/ncpfs/dir.c:136:23: warning: unused variable 'sb' [-Wunused-variable]

This replaces the trivial macro with an equally trivial
inline function, which avoids the warning.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Fixes: 621e155a3591 ("fs: change d_compare for rcu-walk")
---
I sent this on Jan 25 2016 but got no reply, while the warning
remains. Please apply.