[SUBMITTED] ARM: pxa: don't select RFKILL if CONFIG_NET is disabled
Bluetooth is only supported when network support is part of the kernel,
so it is a bit pointless to build the tosa-bt support without networking.
If we try anyway, we get a Kconfig warning:

warning: (TOSA_BT && H1940BT) selects RFKILL which has unmet direct dependencies (NET)

This adds a dependency on CONFIG_NET to avoid that case.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>