Kbuild: shut up unwanted clang warnings
clang warns about inline functions that are defined in a .c file and
never used, but gcc does not warn about that, so we get countless
warnings. The same thing happens when more than one 'const' or other
specifiers appear for a declaration.

This shuts up both kinds of warnings.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>