futex: avoid fixups in __init section
We may get a link error on ARM when we have an SMP fixup in
a .init.text section:

kernel/built-in.o:(.text.fixup+0x206): relocation truncated to fit: R_ARM_THM_JUMP24 against `.init.text'

gcc does not inline cmpxchg_futex_value_locked, and if we prevent
clang from inlining it, everything works.

A better fix would probably be to change the fixup assembly
to be a proper symbol so that ld generates a trampoline.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>