i2c: rework i2cdev_ioctl_smbus handling
The compat handler for I2C_SMBUS copies the user space data into
another user space structure and from there it gets copied to
the kernel. This is unnecessary, and we can simplify it by
copying the data just once and then call a function that takes
a kernel pointer.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>