scsi: sg: move SG_GET_REQUEST_TABLE handler into sg.c
Only the sg driver requires compat ioctl handling of the
SG_GET_REQUEST_TABLE command, so the handler should live
in sg.c, which can be a loadable module or completely
disabled.

Moving the handler there also lets us share the implementation
and avoid a round-trip copying data from user to kernel, back
to user space and again to the kernel.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>