scsi: sg: handle compatible ioctls directly
A number of ioctl commands that the sg.c driver understands are
correctly marked as compatible in fs/compat_ioctl.c and should be
handled by the normal compat_ioctl function, which is more efficient
and straightforward.

This adds all the ones that are listed to sg_compat_ioctl. Unfortunately,
the same commands are handled by a number of other drivers as well,
so we can't easily remove the global definitions.

The SG_IO case is special because it requires going through the
translation in fs/compat_ioctl.c.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>