scsi: sg: add missing compat_ioctl translations
A handful of ioctl commmands that are interpreted by the sg_ioctl
function do not have any kind of handler in fs/compat_ioctl.c.

This adds the missing pieces. Most commands are compatible
and can be added trivially, but the BLKTRACESETUP command needs
a little extra care on x86, where its data structure is slightly
incompatible between 32-bit and 64-bit user space.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>