fs/compat_ioctl.c: remove DVB ioctl conversion functions
A set of helpers for DVB ioctl conversion functions was added over
a decade ago, but they were defined incorrectly and never had any
effect because they match the numbers for the 64-bit ioctl commands,
which are different from the ones used on 32-bit processes because
of the structure size.

I suspect that nobody cared enough about those handlers to notice
that they were not used, so it's easier to just remove them than
to try fixing them properly. If anyone ever wants to do it, the
handlers should be moved into the two drivers that implement the
calls in the first place: ivtv and ttpci.

Of the three calls, only VIDEO_GET_EVENT is actually implemented
by both drivers, while VIDEO_STILLPICTURE is only present in ttpci
and VIDEO_SET_SPU_PALETTE apparently has never been used in an
upstream kernel.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Fixes: 6e87abd0b8cb ("[DVB]: Add compat ioctl handling.")