ppp: move compat ioctl handling into ppp_generic.c
The compat handler for the ppp ioctl commands should not
live in the global fs/compat_ioctl.c file. The only other
user of the translation functions is i4l, and that now has
its own copy, so the code can simply be moved.

All other ioctls that get handled by ppp_generic or one
of the ppp_channel_ops implementation are compatible, so
we can just call ppp_ioctl for those.

However, we cannot simply remove the COMPATIBLE_IOCTL()
annotations in fs/compat_ioctl.c because those commands can
be called either through the ppp device or through a socket
in some cases.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>