isdn: gigaset: remove i4l code
As a rule, drivers that are in staging must not be a dependency
for drivers that are not in staging.

The gigaset driver can be used with either the CAPI or the I4L
interface, so when we move I4L into staging, it is not clear
what should happen to gigaset.

This patch cuts the ties the hard way and removes all i4l
support from gigaset. The driver should keep providing its
full functionality through the capidrv driver in i4l even
when that is moved to staging.

Further cleanup can be done on top of this patch to remove the
abstraction layer in gigaset.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>