Kbuild: allow optimized inlining on ARM
The CONFIG_OPTIMIZE_INLINING option is only available on x86, but
there is not really anything architecture specific in it. According
to 765c68bd54c7 ("generic: make optimized inlining arch-opt-in"),
at least one architecture (powerpc) had problems with the option
before, so we need to be a bit cautious. If problems arise from
setting CONFIG_OPTIMIZE_INLINING=y, that usually means some function
needs to be inlined after all and should be explicitly marked
__always_inline rather than just inline.

On an ARM tinyconfig build, the .text size shrinks by 2.6% with
the option set.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>