[SUBMITTED] staging: amso1100: fix format string for 64-bit phys_addr
The amso1100 driver prints a phys_addr_t by casting it to a pointer,
which causes a warning when phys_addr_t is 64-bit and pointers
are 32-bit:

drivers/staging/rdma/amso1100/c2.c: In function 'c2_rx_error':
drivers/staging/rdma/amso1100/c2.c:430:4: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]

This changes the format string to print the address as a 64-bit
number all the time to avoids the warnings. Ideally we would use
the %pap format string for phys_addr_t, but that doesn't work
here easily as it requires passing a pointer to the address.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>